Add OptionalTensor type
This commit adds an OptionalTensor type that allows for the highlevel
`Maybe Tensor` type to be be converted to `std::optional`.

The motivating function for this feature is
`scaled_dot_product_attention` which will throw a runtime error if
both `attn_mask` and `is_causal` is set to true. Without
OptionalTensor, there is no way to specify `std::nullopt` (or None
essentially) for the attention mask and hence `is_casusal` can never
be set to true. More information can be found here in the torch
documentation
https://docs.pytorch.org/docs/stable/generated/torch.nn.functional.scaled_dot_product_attention.html#torch.nn.functional.scaled_dot_product_attention